fix: add support to create IPv6 clusters
Before this change, there was not a way to set the IpFamily for a
cluster. This adds the ability to set IpFamily based upon a naming
scheme where the cluster name ends with ipv6. This adds a new
way of creating a cluster by building the eksctl config file before
creating the cluster.